refactor(ipc): split repos.ts into focused modules
