+++ src/main/ipc/folder-repo-git-upgrade.ts
@@ -12,7 +12,7 @@ import { isWslUncPath } from '../../shared/wsl-paths'
 import { getGitRepoRoot, isGitRepo } from '../git/repo'
 import { prepareLocalWorktreeRootForRepo } from '../worktree-root-preparation'
 import { invalidateAuthorizedRootsCache } from './registered-worktree-roots-cache'
-import { notifyReposChanged } from './repos'
+import { notifyReposChanged } from './repos/repos-changed-notification'
 import { notifyWorktreesChanged } from './worktree-remote'
 import { setFolderRepoGitUpgradeWakeListener } from './folder-repo-git-upgrade-wake'
 import {
