+++ src/main/ipc/repos/nested-repo-import-handler.ts
@@ -1,0 +1,185 @@
+import type { BrowserWindow } from 'electron'
+import { ipcMain } from 'electron'
+import { randomUUID } from 'node:crypto'
+import type { Store } from '../../persistence'
+import type { Repo } from '../../../shared/repo-types'
+import type { ProjectGroupImportResult } from '../../../shared/project-group-types'
+import { DEFAULT_REPO_BADGE_COLOR } from '../../../shared/constants'
+import { normalizeRuntimePathForComparison } from '../../../shared/cross-platform-path'
+import { awaitWindowsHostGitEnvironmentReady } from '../../git/runner'
+import { isGitRepo, getRepoName } from '../../git/repo'
+import {
+  createNestedProjectGroupResolver,
+  resolveNestedRepoSelection
+} from '../../project-groups/nested-repo-import'
+import { createNestedRepoImportTargetResolver } from '../../project-groups/nested-repo-import-target'
+import { getSshGitProvider } from '../../providers/ssh-git-dispatch'
+import { detectRepoIconAndUpstream } from '../../repo-icon-autodetect'
+import { prepareLocalWorktreeRootForRepo } from '../../worktree-root-preparation'
+import { getActiveMultiplexer } from '../ssh'
+import { invalidateAuthorizedRootsCache } from '../registered-worktree-roots-cache'
+import { emitRepoAdded } from './repo-added-telemetry'
+import { notifyReposChanged } from './repos-changed-notification'
+import { ProjectGroupImportNestedArgs, parseProjectGroupIpcArgs } from './repo-ipc-arg-schemas'
+import { getCompletedNestedRepoScan, scanNestedReposForIpc } from './nested-repo-scan-ipc'
+
+function sanitizeNestedRepoImportError(context: string, error: unknown): string {
+  console.warn(`[project-groups] ${context}`, error)
+  return 'Repository could not be imported'
+}
+
+export function registerNestedRepoImportHandler(mainWindow: BrowserWindow, store: Store): void {
+  ipcMain.handle(
+    'projectGroups:importNested',
+    async (_event, rawArgs: unknown): Promise<ProjectGroupImportResult> => {
+      const args = parseProjectGroupIpcArgs(
+        ProjectGroupImportNestedArgs,
+        rawArgs,
+        'invalid_project_group_import_nested_args'
+      )
+      const requestedPaths = args.projectPaths
+      const completedScan = getCompletedNestedRepoScan(args)
+      const scan =
+        completedScan ??
+        (await scanNestedReposForIpc({
+          path: args.parentPath,
+          connectionId: args.connectionId,
+          options: { timeoutMs: 15_000 }
+        }))
+      const selection = resolveNestedRepoSelection({ scan, projectPaths: requestedPaths })
+      const groupResolver = createNestedProjectGroupResolver({
+        parentPath: scan.selectedPath,
+        groupName: args.groupName ?? '',
+        mode: args.mode,
+        connectionId: args.connectionId ?? null,
+        repoPaths: selection.selectedPaths,
+        createGroup: (input) => store.createProjectGroup(input)
+      })
+      const results: ProjectGroupImportResult['projects'] = selection.rejectedPaths.map(
+        (repoPath) => ({
+          path: repoPath,
+          status: 'failed',
+          error: 'Repository was not found in the nested repo scan result'
+        })
+      )
+      const importedProjectIdsByRepoPath = new Map<string, string>()
+      const importTargetResolver = createNestedRepoImportTargetResolver()
+
+      for (const [projectGroupOrder, repoPath] of selection.selectedPaths.entries()) {
+        try {
+          let importRepoPath = repoPath
+          if (args.connectionId) {
+            const gitProvider = getSshGitProvider(args.connectionId)
+            const check = gitProvider ? await gitProvider.isGitRepoAsync(repoPath) : null
+            if (!gitProvider || !check?.isRepo) {
+              results.push({
+                path: repoPath,
+                status: 'failed',
+                error: 'Not a valid git repository'
+              })
+              continue
+            }
+            importRepoPath = await importTargetResolver.resolveSsh(repoPath, gitProvider)
+          } else {
+            await awaitWindowsHostGitEnvironmentReady({ cwd: repoPath })
+            if (!isGitRepo(repoPath)) {
+              results.push({
+                path: repoPath,
+                status: 'failed',
+                error: 'Not a valid git repository'
+              })
+              continue
+            }
+            importRepoPath = await importTargetResolver.resolveLocal(repoPath)
+          }
+          const normalizedImportRepoPath = normalizeRuntimePathForComparison(importRepoPath)
+          const alreadyImportedProjectId =
+            importedProjectIdsByRepoPath.get(normalizedImportRepoPath)
+          if (alreadyImportedProjectId) {
+            results.push({
+              path: repoPath,
+              projectId: alreadyImportedProjectId,
+              status: 'already-known'
+            })
+            continue
+          }
+          const existing = store
+            .getRepos()
+            .find(
+              (repo) =>
+                (repo.connectionId ?? null) === (args.connectionId ?? null) &&
+                normalizeRuntimePathForComparison(repo.path) === normalizedImportRepoPath
+            )
+          const group = groupResolver.getGroupForRepo(repoPath)
+          if (existing) {
+            if (group) {
+              store.moveProjectToGroup(existing.id, group.id, projectGroupOrder)
+            }
+            importedProjectIdsByRepoPath.set(normalizedImportRepoPath, existing.id)
+            results.push({ path: repoPath, projectId: existing.id, status: 'already-known' })
+            continue
+          }
+          const detected = await detectRepoIconAndUpstream({
+            repoPath: importRepoPath,
+            kind: 'git',
+            connectionId: args.connectionId
+          })
+          const repo: Repo = {
+            id: randomUUID(),
+            path: importRepoPath,
+            displayName: getRepoName(importRepoPath),
+            badgeColor: DEFAULT_REPO_BADGE_COLOR,
+            ...detected,
+            addedAt: Date.now(),
+            kind: 'git',
+            ...(args.connectionId ? { connectionId: args.connectionId } : {}),
+            externalWorktreeVisibilityLegacy: false,
+            projectHostSetupMethod: 'imported-existing-folder',
+            ...(group
+              ? {
+                  projectGroupId: group.id,
+                  projectGroupOrder
+                }
+              : {})
+          }
+          store.addRepo(repo)
+          await prepareLocalWorktreeRootForRepo(store, repo)
+          if (args.connectionId) {
+            getActiveMultiplexer(args.connectionId)?.notify('session.registerRoot', {
+              rootPath: importRepoPath
+            })
+          }
+          importedProjectIdsByRepoPath.set(normalizedImportRepoPath, repo.id)
+          results.push({ path: repoPath, projectId: repo.id, status: 'imported' })
+          // Why: reaches here only after the isGitRepo guard above confirmed a git repo, so always true.
+          emitRepoAdded('folder_picker', false, true)
+        } catch (error) {
+          results.push({
+            path: repoPath,
+            status: 'failed',
+            error: sanitizeNestedRepoImportError('Failed to import nested repository', error)
+          })
+        }
+      }
+
+      const importedCount = results.filter((entry) => entry.status === 'imported').length
+      const alreadyKnownCount = results.filter((entry) => entry.status === 'already-known').length
+      const failedCount = results.filter((entry) => entry.status === 'failed').length
+      if (importedCount + alreadyKnownCount === 0) {
+        for (const group of groupResolver.getCreatedGroups().toReversed()) {
+          store.deleteProjectGroup(group.id)
+        }
+      }
+      invalidateAuthorizedRootsCache()
+      notifyReposChanged(mainWindow)
+      const rootGroup = groupResolver.getRootGroup()
+      return {
+        ...(rootGroup && importedCount + alreadyKnownCount > 0 ? { group: rootGroup } : {}),
+        projects: results,
+        importedCount,
+        alreadyKnownCount,
+        failedCount
+      }
+    }
+  )
+}
