+++ src/main/ipc/repos/remote-home-path.ts
@@ -1,0 +1,18 @@
+import { getActiveMultiplexer } from '../ssh'
+
+export async function resolveRemoteHomePath(connectionId: string, path: string): Promise<string> {
+  if (path !== '~' && path !== '~/' && !path.startsWith('~/')) {
+    return path
+  }
+  const mux = getActiveMultiplexer(connectionId)
+  if (!mux) {
+    return path
+  }
+  try {
+    const result = (await mux.request('session.resolveHome', { path })) as { resolvedPath: string }
+    return result.resolvedPath
+  } catch {
+    // Why: older relays may not support this; return the original path so callers surface their own validation error.
+    return path
+  }
+}
