+++ src/main/ipc/repos/repo-clone-lifecycle.ts
@@ -1,0 +1,299 @@
+import type { BrowserWindow } from 'electron'
+import { ipcMain } from 'electron'
+import type { ChildProcess } from 'node:child_process'
+import { randomUUID } from 'node:crypto'
+import { mkdir } from 'node:fs/promises'
+import type { Store } from '../../persistence'
+import type { Repo } from '../../../shared/repo-types'
+import { isFolderRepo } from '../../../shared/repo-kind'
+import { DEFAULT_REPO_BADGE_COLOR } from '../../../shared/constants'
+import { getGitCloneFailureMessage } from '../../../shared/git-clone-failure-message'
+import { gitSpawnAfterWindowsEnvironmentReady, nonInteractiveGitEnv } from '../../git/runner'
+import { getRepoName } from '../../git/repo'
+import type { ClaimedCloneTarget } from '../../git/repo-clone-path'
+import {
+  cleanupClaimedCloneTarget,
+  claimCloneTarget,
+  deriveValidatedClonePath,
+  getClonePathComparisonKey
+} from '../../git/repo-clone-path'
+import { detectRepoIconAndUpstream } from '../../repo-icon-autodetect'
+import { prepareLocalWorktreeRootForRepo } from '../../worktree-root-preparation'
+import { invalidateAuthorizedRootsCache } from '../registered-worktree-roots-cache'
+import { emitRepoAdded } from './repo-added-telemetry'
+import { notifyReposChanged } from './repos-changed-notification'
+import { runWithClonePathLock } from './clone-path-lock'
+import { abortActiveRemoteClone, cloneRemoteRepo } from './remote-repo-clone'
+
+type ActiveCloneMetadata = {
+  path: string
+  pathKey: string
+  claimedTarget: ClaimedCloneTarget
+  process: ChildProcess
+  abortRequested: boolean
+  generation: number
+  pendingAbortCleanup: Promise<void> | null
+  resolvePendingAbortCleanup: (() => void) | null
+}
+
+// Why: module-scoped so the abort handle survives macOS window re-creation, when registerRepoHandlers re-runs.
+let activeClone: ActiveCloneMetadata | null = null
+const pendingLocalCloneControllers = new Set<AbortController>()
+let nextCloneGeneration = 1
+const latestCloneGenerationByPath = new Map<string, number>()
+const pendingAbortCleanupByPath = new Map<string, Promise<void>>()
+
+function emitCloneProgressFromText(mainWindow: BrowserWindow, text: string): void {
+  for (const line of text.split(/[\r\n]+/)) {
+    const match = line.match(/^([\w\s]+):\s+(\d+)%/)
+    if (match && !mainWindow.isDestroyed()) {
+      mainWindow.webContents.send('repos:clone-progress', {
+        phase: match[1].trim(),
+        percent: Number.parseInt(match[2], 10)
+      })
+    }
+  }
+}
+
+async function cleanupOwnedCloneTarget(metadata: ActiveCloneMetadata): Promise<void> {
+  if (!metadata.claimedTarget.canCleanup || !metadata.claimedTarget.ownedDirectoryIdentity) {
+    return
+  }
+  if (latestCloneGenerationByPath.get(metadata.pathKey) !== metadata.generation) {
+    return
+  }
+  // Why: a fast retry may attach a newer process before the aborted one closes; the old close handler must not delete it.
+  if (
+    activeClone &&
+    activeClone.process !== metadata.process &&
+    activeClone.pathKey === metadata.pathKey
+  ) {
+    return
+  }
+
+  if (latestCloneGenerationByPath.get(metadata.pathKey) !== metadata.generation) {
+    return
+  }
+  await cleanupClaimedCloneTarget(metadata.path, metadata.claimedTarget)
+}
+
+function markCloneAbortCleanupPending(metadata: ActiveCloneMetadata): void {
+  if (metadata.resolvePendingAbortCleanup) {
+    return
+  }
+  metadata.pendingAbortCleanup = new Promise<void>((resolve) => {
+    metadata.resolvePendingAbortCleanup = resolve
+  })
+  pendingAbortCleanupByPath.set(metadata.pathKey, metadata.pendingAbortCleanup)
+}
+
+function settleCloneAbortCleanup(metadata: ActiveCloneMetadata): void {
+  if (pendingAbortCleanupByPath.get(metadata.pathKey) === metadata.pendingAbortCleanup) {
+    pendingAbortCleanupByPath.delete(metadata.pathKey)
+  }
+  metadata.resolvePendingAbortCleanup?.()
+  metadata.pendingAbortCleanup = null
+  metadata.resolvePendingAbortCleanup = null
+}
+
+export function registerRepoCloneHandlers(mainWindow: BrowserWindow, store: Store): void {
+  ipcMain.handle('repos:cloneAbort', async () => {
+    for (const controller of pendingLocalCloneControllers) {
+      controller.abort()
+    }
+    pendingLocalCloneControllers.clear()
+    if (activeClone) {
+      const clone = activeClone
+      clone.abortRequested = true
+      markCloneAbortCleanupPending(clone)
+      clone.process.kill()
+      activeClone = null
+    }
+    abortActiveRemoteClone()
+  })
+
+  ipcMain.handle(
+    'repos:clone',
+    async (_event, args: { url: string; destination: string }): Promise<Repo> => {
+      // Why: derive the repo folder name from the URL's last segment, matching default git clone behavior.
+      const clonePath = deriveValidatedClonePath(args)
+      const clonePathKey = getClonePathComparisonKey(clonePath)
+      return runWithClonePathLock(clonePathKey, async () => {
+        await pendingAbortCleanupByPath.get(clonePathKey)
+        const existingAfterPendingClone = store
+          .getRepos()
+          .find((r) => getClonePathComparisonKey(r.path) === clonePathKey)
+        if (existingAfterPendingClone && !isFolderRepo(existingAfterPendingClone)) {
+          // Why: clone_url always produces a git repo.
+          emitRepoAdded('clone_url', true, true)
+          return existingAfterPendingClone
+        }
+        // Why: gitSpawn cwd is args.destination, so it must exist before spawn (fresh installs may lack the defaulted parent).
+        await mkdir(args.destination, { recursive: true })
+        const claimedTarget = await claimCloneTarget(clonePath)
+
+        // Why: spawn (not execFile) avoids the maxBuffer limit — clone progress on stderr can exceed Node's 1 MB default.
+        // Why: --progress forces git to emit progress even when stderr isn't a TTY.
+        const cloneMetadataRef: { current: ActiveCloneMetadata | null } = { current: null }
+        let proc: Awaited<ReturnType<typeof gitSpawnAfterWindowsEnvironmentReady>>
+        const pendingController = new AbortController()
+        pendingLocalCloneControllers.add(pendingController)
+        try {
+          // Why: use the parent destination as cwd so the runner detects a WSL path and routes through wsl.exe.
+          // Why: '--' isolates the URL so a malicious URL can't be read as git flags (command injection).
+          proc = await gitSpawnAfterWindowsEnvironmentReady(
+            ['clone', '--progress', '--', args.url, clonePath],
+            {
+              cwd: args.destination,
+              // Why: without this, an auth-needing clone pops Git Credential Manager's OAuth window on Windows, unclosable in a restricted env (issue #7652).
+              env: nonInteractiveGitEnv(),
+              signal: pendingController.signal,
+              stdio: ['ignore', 'ignore', 'pipe']
+            }
+          )
+        } catch (err) {
+          await cleanupClaimedCloneTarget(clonePath, claimedTarget)
+          const message = err instanceof Error ? err.message : String(err)
+          throw new Error(`Clone failed: ${message}`)
+        } finally {
+          pendingLocalCloneControllers.delete(pendingController)
+        }
+        await new Promise<void>((resolve, reject) => {
+          const generation = nextCloneGeneration++
+          latestCloneGenerationByPath.set(clonePathKey, generation)
+          const metadata: ActiveCloneMetadata = {
+            path: clonePath,
+            pathKey: clonePathKey,
+            claimedTarget,
+            process: proc,
+            abortRequested: false,
+            generation,
+            pendingAbortCleanup: null,
+            resolvePendingAbortCleanup: null
+          }
+          cloneMetadataRef.current = metadata
+          activeClone = metadata
+
+          let stderrTail = ''
+          let settled = false
+          proc.stderr!.on('data', (chunk: Buffer) => {
+            const text = chunk.toString()
+            stderrTail = (stderrTail + text).slice(-4096)
+
+            // Why: git progress lines use \r to overwrite in-place; parse fragments the same as SSH clone.
+            emitCloneProgressFromText(mainWindow, text)
+          })
+
+          const finishClone = async (
+            code: number | null,
+            signal: NodeJS.Signals | null,
+            err?: Error
+          ) => {
+            if (settled) {
+              return
+            }
+            settled = true
+            // Why: only null activeClone if it still points to this proc; abort-and-retry may have reassigned it, stranding the new clone.
+            if (activeClone?.process === proc) {
+              activeClone = null
+            }
+
+            const cloneSucceeded = !err && code === 0 && !signal
+            if (!cloneSucceeded) {
+              // Why: only the process that created this target may remove it, and only after git reports failure.
+              await cleanupOwnedCloneTarget(metadata)
+            }
+            if (metadata.abortRequested && !cloneSucceeded) {
+              settleCloneAbortCleanup(metadata)
+            }
+            if (latestCloneGenerationByPath.get(metadata.pathKey) === metadata.generation) {
+              latestCloneGenerationByPath.delete(metadata.pathKey)
+            }
+
+            if (err) {
+              reject(new Error(`Clone failed: ${err.message}`))
+            } else if (signal === 'SIGTERM') {
+              reject(new Error('Clone aborted'))
+            } else if (code === 0) {
+              resolve()
+            } else {
+              reject(
+                new Error(`Clone failed: ${getGitCloneFailureMessage(stderrTail, { clonePath })}`)
+              )
+            }
+          }
+
+          proc.on('error', (err) => {
+            void finishClone(null, null, err)
+          })
+
+          proc.on('close', (code, signal) => {
+            void finishClone(code, signal)
+          })
+        })
+
+        try {
+          // Why: check after clone (path didn't exist before); reuse+upgrade a folder repo clone landed into instead of duplicating.
+          const existing = store
+            .getRepos()
+            .find((r) => getClonePathComparisonKey(r.path) === clonePathKey)
+          if (existing) {
+            if (isFolderRepo(existing)) {
+              const updated = store.updateRepo(existing.id, {
+                kind: 'git',
+                projectHostSetupMethod: 'cloned'
+              })
+              if (updated) {
+                await prepareLocalWorktreeRootForRepo(store, updated)
+                invalidateAuthorizedRootsCache()
+                notifyReposChanged(mainWindow)
+                // Why: folder→git upgrade is a real new git repo provisioning event.
+                emitRepoAdded('clone_url', false, true)
+                return updated
+              }
+            }
+            emitRepoAdded('clone_url', true, true)
+            return existing
+          }
+
+          const detected = await detectRepoIconAndUpstream({ repoPath: clonePath, kind: 'git' })
+          const repo: Repo = {
+            id: randomUUID(),
+            path: clonePath,
+            displayName: getRepoName(clonePath),
+            badgeColor: DEFAULT_REPO_BADGE_COLOR,
+            ...detected,
+            addedAt: Date.now(),
+            kind: 'git',
+            externalWorktreeVisibilityLegacy: false,
+            projectHostSetupMethod: 'cloned'
+          }
+
+          store.addRepo(repo)
+          await prepareLocalWorktreeRootForRepo(store, repo)
+          invalidateAuthorizedRootsCache()
+          notifyReposChanged(mainWindow)
+          emitRepoAdded('clone_url', false, true)
+          return repo
+        } finally {
+          const metadata = cloneMetadataRef.current
+          if (metadata?.abortRequested) {
+            settleCloneAbortCleanup(metadata)
+          }
+        }
+      })
+    }
+  )
+
+  ipcMain.handle(
+    'repos:cloneRemote',
+    async (
+      _event,
+      args: { connectionId: string; url: string; destination: string }
+    ): Promise<Repo> => {
+      const repo = await cloneRemoteRepo(store, mainWindow, args)
+      notifyReposChanged(mainWindow)
+      return repo
+    }
+  )
+}
