+++ src/main/ipc/repos/sparse-preset-handlers.ts
@@ -1,0 +1,90 @@
+import type { BrowserWindow } from 'electron'
+import { ipcMain } from 'electron'
+import { randomUUID } from 'node:crypto'
+import type { Store } from '../../persistence'
+import type { SparsePreset } from '../../../shared/worktree/create-types'
+import { normalizeSparseDirectories } from '../sparse-checkout-directories'
+
+export function registerSparsePresetHandlers(mainWindow: BrowserWindow, store: Store): void {
+  // ── Sparse presets ─────────────────────────────────────────────
+  // Why: repo-scoped reusable directory lists for the new-workspace composer; broadcast on change so open composers refresh.
+
+  ipcMain.handle('sparsePresets:list', (_event, args: { repoId: string }) => {
+    return store.getSparsePresets(args.repoId)
+  })
+
+  ipcMain.handle(
+    'sparsePresets:save',
+    (
+      _event,
+      args: { repoId: string; id?: string; name: string; directories: string[] }
+    ): SparsePreset => {
+      const repo = store.getRepo(args.repoId)
+      if (!repo) {
+        throw new Error(`Repo "${args.repoId}" not found`)
+      }
+      const name = normalizeSparsePresetName(args.name)
+      const directories = normalizeSparsePresetDirectories(args.directories)
+      const now = Date.now()
+      const existing = args.id
+        ? store.getSparsePresets(args.repoId).find((preset) => preset.id === args.id)
+        : undefined
+      const preset: SparsePreset = {
+        id: existing?.id ?? randomUUID(),
+        repoId: args.repoId,
+        name,
+        directories,
+        createdAt: existing?.createdAt ?? now,
+        updatedAt: now
+      }
+      const saved = store.saveSparsePreset(preset)
+      notifySparsePresetsChanged(mainWindow, args.repoId)
+      return saved
+    }
+  )
+
+  ipcMain.handle('sparsePresets:remove', (_event, args: { repoId: string; presetId: string }) => {
+    const repo = store.getRepo(args.repoId)
+    if (!repo) {
+      throw new Error(`Repo "${args.repoId}" not found`)
+    }
+    store.removeSparsePreset(args.repoId, args.presetId)
+    notifySparsePresetsChanged(mainWindow, args.repoId)
+  })
+}
+
+function notifySparsePresetsChanged(mainWindow: BrowserWindow, repoId: string): void {
+  if (!mainWindow.isDestroyed()) {
+    mainWindow.webContents.send('sparsePresets:changed', { repoId })
+  }
+}
+
+function normalizeSparsePresetName(name: string): string {
+  const trimmed = name.trim()
+  if (!trimmed) {
+    throw new Error('Preset name is required.')
+  }
+  if (trimmed.length > 80) {
+    throw new Error('Preset name is too long.')
+  }
+  return trimmed
+}
+
+function normalizeSparsePresetDirectories(directories: string[]): string[] {
+  let normalized: string[]
+  try {
+    normalized = normalizeSparseDirectories(directories)
+  } catch (err) {
+    if (
+      err instanceof Error &&
+      err.message === 'Sparse checkout directories must be repo-relative paths.'
+    ) {
+      throw new Error('Preset directories must be repo-relative paths.')
+    }
+    throw err
+  }
+  if (normalized.length === 0) {
+    throw new Error('Preset must have at least one directory.')
+  }
+  return normalized
+}
