+++ src/main/window/attach-main-window-services.ts
@@ -14,7 +14,8 @@ import {
   dismissTccPromptNotice,
   releasePendingTccPromptNotice
 } from '../macos-tcc-prompt-notice'
-import { registerRepoHandlers, setRepoRemoteClientNotifier } from '../ipc/repos'
+import { registerRepoHandlers } from '../ipc/repos'
+import { setRepoRemoteClientNotifier } from '../ipc/repos/repos-changed-notification'
 import { registerWorktreeHandlers } from '../ipc/worktrees'
 import { registerWorkspaceCleanupHandlers } from '../ipc/workspace-cleanup'
 import {
